third_party/WebKit/LayoutTests/typedcssom/TransformValue.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>

test(function() {
	var transformValueObject = new TransformValue();
    assert_true(transformValueObject instanceof TransformValue);
    assert_true(transformValueObject instanceof StyleValue);
}, "A TransformValue object can be constructed");

</script>

<body>
</body>
